requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.9
sentence-transformers==2.3.1
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.31.0
celery==5.3.6
redis==5.0.1
numpy==1.26.4
torch==2.2.1
